part1end

+++ libft/ft_strncmp.c
@@ -6,7 +6,7 @@
 /*   By: sunmin <msh4287@naver.com>                 +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/03 16:32:44 by sunmin            #+#    #+#             */
-/*   Updated: 2020/10/03 16:32:45 by sunmin           ###   ########.fr       */
+/*   Updated: 2020/10/06 10:46:06 by sunmin           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strncmp(const char *s1, const char *s2, size_t n)
 {
-	int i;
+	unsigned long i;
 
 	i = 0;
 	while (s1[i] && s2[i] && i < n)
